// src/app/generated/dcc/model/byteDataDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { RichContentDto } from './richContentDto';


export interface ByteDataDto { 
    id?: string;
    refIds?: Array<string>;
    name?: LanguageSpecificStringsDto;
    content?: string;
    mimeType?: string;
    fileName?: string;
    description?: RichContentDto;
}