// src/app/generated/dcc/model/locationDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RichContentDto } from './richContentDto';


export interface LocationDto { 
    id?: string;
    countryCode?: string;
    stateCode?: string;
    city?: string;
    postalCode?: string;
    street?: string;
    houseNumber?: string;
    poBox?: string;
    additionalInformation?: RichContentDto;
}